# libs/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/workflows": {
      "get": {
        "operationId": "WorkflowsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "ステータス",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "submitted",
                "approved",
                "rejected",
                "withdrawn"
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "申請タイプ",
            "schema": {
              "type": "string",
              "enum": [
                "expense",
                "leave",
                "purchase",
                "other"
              ]
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "description": "モード",
            "schema": {
              "type": "string",
              "enum": [
                "mine",
                "pending"
              ]
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "開始日",
            "schema": {
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "終了日",
            "schema": {
              "example": "2026-01-31",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "ページ番号",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "1ページの件数",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "WorkflowsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/pending": {
      "get": {
        "operationId": "WorkflowsController_findPending",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{id}": {
      "get": {
        "operationId": "WorkflowsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      },
      "patch": {
        "operationId": "WorkflowsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{id}/submit": {
      "post": {
        "operationId": "WorkflowsController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{id}/approve": {
      "post": {
        "operationId": "WorkflowsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{id}/reject": {
      "post": {
        "operationId": "WorkflowsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{id}/withdraw": {
      "post": {
        "operationId": "WorkflowsController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{id}/attachments": {
      "post": {
        "operationId": "WorkflowsController_uploadAttachment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "operationId": "WorkflowsController_getAttachments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{id}/attachments/{attachmentId}": {
      "delete": {
        "operationId": "WorkflowsController_deleteAttachment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{id}/attachments/{attachmentId}/download": {
      "get": {
        "operationId": "WorkflowsController_downloadAttachment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workflows"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpsHub API",
    "description": "OpsHub バックエンド API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateWorkflowDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "申請タイプ",
            "enum": [
              "expense",
              "leave",
              "purchase",
              "other"
            ]
          },
          "title": {
            "type": "string",
            "description": "タイトル",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "説明",
            "maxLength": 2000
          },
          "amount": {
            "type": "number",
            "description": "金額"
          },
          "dateFrom": {
            "type": "string",
            "description": "開始日",
            "example": "2026-01-01"
          },
          "dateTo": {
            "type": "string",
            "description": "終了日",
            "example": "2026-01-31"
          },
          "approverId": {
            "type": "string",
            "description": "承認者ID（UUID）"
          },
          "action": {
            "type": "string",
            "description": "アクション",
            "enum": [
              "draft",
              "submit"
            ]
          }
        },
        "required": [
          "type",
          "title",
          "approverId",
          "action"
        ]
      },
      "UpdateWorkflowDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "タイトル",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "説明",
            "maxLength": 2000
          },
          "amount": {
            "type": "number",
            "description": "金額"
          },
          "dateFrom": {
            "type": "string",
            "description": "開始日",
            "example": "2026-01-01"
          },
          "dateTo": {
            "type": "string",
            "description": "終了日",
            "example": "2026-01-31"
          },
          "approverId": {
            "type": "string",
            "description": "承認者ID（UUID）"
          }
        }
      },
      "RejectWorkflowDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "差戻し理由"
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}